+++ BackEnd/src/main/java/com/MapView/BackEnd/repository/EquipmentRepository.java
@@ -15,7 +15,7 @@ public interface EquipmentRepository extends JpaRepository<Equipment,String> {
 
     Optional<Equipment> findByValidity(String validity);
     List<Equipment> findByModel(EnumModelEquipment equipment);
-    Equipment findByRfid(int rfid);
+    Equipment findByRfid(Long rfid);
 
     //Equipment findByRfid(String rfid);
 }
